+++ src/configs/mobx.ts
@@ -7,10 +7,10 @@ const config: Array<Linter.Config> = [
   {
     name: '@dartess/mobx',
     rules: {
-      'mobx/missing-observer': 'off', // replaced by the neater "@dartess/require-observer"
+      'mobx/missing-observer': 'off', // replaced by the neater "@dartess/mobx-require-observer"
       'mobx/missing-make-observable': 'off', // useless with modern decorators syntax. TODO check original plugin?
-      '@dartess/strict-observable-components-declaration': 'error',
-      '@dartess/require-observer': 'error',
+      '@dartess/mobx-strict-observable-components-declaration': 'error',
+      '@dartess/mobx-require-observer': 'error',
       '@dartess/mobx-sync-autorun': 'error', // TODO implement it by types?
       '@dartess/mobx-sync-action': 'error', // TODO implement it by types?
     },